+++ public/js/factories/Topic.js
@@ -124,6 +124,28 @@ angular
                             return angular.fromJson(data);
                         }
                     }
+                },
+                addToFavourites: {
+                    method: 'POST',
+                    params: {topicId: '@id'},
+                    url: sLocation.getAbsoluteUrlApi('/api/users/self/topics/:topicId/favourites'),
+                    transformResponse: function (data, headersGetter, status) {
+                        if (status > 0 && status < 400) { // IF patch is working then make it return data again, for now return nothing to stop from overwriting all fields but topkenJoin
+                        } else {
+                            return angular.fromJson(data);
+                        }
+                    }
+                },
+                removeFromFavourites: {
+                    method: 'DELETE',
+                    params: {topicId: '@id'},
+                    url: sLocation.getAbsoluteUrlApi('/api/users/self/topics/:topicId/favourites'),
+                    transformResponse: function (data, headersGetter, status) {
+                        if (status > 0 && status < 400) { // IF patch is working then make it return data again, for now return nothing to stop from overwriting all fields but topkenJoin
+                        } else {
+                            return angular.fromJson(data);
+                        }
+                    }
                 }
             }
         );